requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.10.0
openai>=0.27.0
numpy>=1.21.0
scipy>=1.7.0
Pillow==10.2.0  # Python Imaging Library for image processing
playwright>=1.41.0  # For web automation and screenshots 